Refactor some controls, and debug
