+++ catkin_ws/src/thruster_control/launch/v1_control.launch
@@ -1,6 +1,6 @@
 <launch>
 
-    <node name="sub_controller" pkg="thruster_control" type="v1_controller.py" output="screen"/>
+    <node name="command_controller" pkg="thruster_control" type="v1_command_manager.py" output="screen"/>
     <node name="thruster_controller" pkg="thruster_control" type="sim_thruster_controller.py" output="screen"/>
     <node name="depth_pid" pkg="thruster_control" type="demo_depth_control.py"/>
