+++ catkin_ws/src/thruster_control/src/thruster_control/movement_command_collector.py
@@ -3,9 +3,6 @@
 import rospy
 import numpy as np
 
-from thruster_control.configurations.v2_configuration import get_wrench_to_thrusts_lb_in
-
-from aquadrone_msgs.msg import MotorControls
 from geometry_msgs.msg import Wrench
 
 
@@ -23,27 +20,20 @@ class CommandSubscriber:
         return self.cmd, (rospy.Time() - self.time).to_sec()
 
 
-class WrenchConverter:
-
-    def __init__(self, w_to_t):
-        self.transform = w_to_t
-
+class MovementCommandCollector:
+    def __init__(self):
+        # First added is highest priority
         self.sources = []
         self.sources.append(CommandSubscriber("movement_command"))
         self.sources.append(CommandSubscriber("depth_command"))
 
-        self.publisher = rospy.Publisher("motor_command", MotorControls, queue_size=0)
+    def get_recent_thrusts(self, drop=0):
+        w = self.get_empty_wrench()
 
-    def run(self):
-        while not rospy.is_shutdown():
-            w = self.get_empty_thrusts()
+        for i in range(0, len(self.sources) - drop):
+            w = w + self.add_source_command(self.sources[i])
 
-            for source in self.sources:
-                w = w + self.add_source_command(source)
-
-            self.publish_command(w)
-
-            rospy.sleep(0.1)
+        return w
 
     def add_source_command(self, source):
         cmd, dt = source.get_cmd()
@@ -51,20 +41,11 @@ class WrenchConverter:
             cmd = cmd * 0.0
         return self.wrench_to_np_array(cmd)
 
-
-    def publish_command(self, force):
-        thrusts = np.dot(self.transform, force)
-        msg = MotorControls()
-
-        th = np.ndarray.tolist(thrusts)[0]
-        print(th)
-        msg.motorThrusts = list(th)
-        self.publisher.publish(msg)
-
-    def get_empty_thrusts(self):
+    def get_empty_wrench(self):
         return self.wrench_to_np_array(Wrench())
     
     def wrench_to_np_array(self, wrench):
         return np.array([wrench.force.x, wrench.force.y, wrench.force.z, 
                          wrench.torque.x, wrench.torque.y, wrench.torque.z])
+
